pyproject.toml
[project]
name = "paper-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    # "onnxruntime-gpu>=1.19.2",
    "openai>=1.52.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
find-links = [
    "https://download.pytorch.org/whl/cu121/torch",
]
